% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mori-package.R
\docType{package}
\name{mori-package}
\alias{mori}
\alias{mori-package}
\title{mori: Shared Memory for R Objects}
\description{
Share R objects via shared memory with \code{\link[=share]{share()}}, access them in
other processes with \code{\link[=map_shared]{map_shared()}}, using R's ALTREP framework for
zero-copy memory-mapped access. Shared objects serialize compactly
via ALTREP serialization hooks. Shared memory is automatically freed
when the R object is garbage collected.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-lib.github.io/mori/}
  \item \url{https://github.com/r-lib/mori}
  \item Report bugs at \url{https://github.com/r-lib/mori/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Authors:
\itemize{
  \item Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
